Populate bambu config folder if non-existent

+++ Install Printer Profiles.ps1
@@ -1,11 +1,48 @@
+# see https://stackoverflow.com/a/28237896
+Function pause ($message)
+{
+    # Check if running Powershell ISE
+    if ($psISE)
+    {
+        Add-Type -AssemblyName System.Windows.Forms
+        [System.Windows.Forms.MessageBox]::Show("$message")
+    }
+    else
+    {
+        Write-Host "$message" -BackgroundColor Yellow -ForegroundColor Black
+        $x = $host.ui.RawUI.ReadKey("NoEcho,IncludeKeyDown")
+    }
+}
+
 # Download and install PrusaSlicer printer profiles
+Write-Output "Downloading and installing PrusaSlicer profiles..."
 wget https://github.com/democat3457/PrusaProfiles/raw/refs/heads/main/config.zip -OutFile config.zip
 Expand-Archive -Force config.zip -DestinationPath $env:APPDATA\PrusaSlicer\printer
 rm config.zip
+Write-Output "PrusaSlicer profile installation complete!"
 
-# Download and update BambuSlicer printer access codes
+# Check if Bambu Studio has started
+Write-Output "Downloading and installing Bambu Studio printer information..."
+if (!(Test-Path "$($env:APPDATA)\BambuStudio" -PathType Container)) {
+    Write-Host "Bambu Studio config folder not found! This script will open Bambu Studio to generate the config folder, then automatically close it and continue with the script."
+    pause "Press any key to continue..."
+    Start-Process -NoNewWindow -FilePath 'C:\Program Files\Bambu Studio\bambu-studio.exe'
+    Start-Sleep -Seconds 12
+    Stop-Process -Name "bambu-studio"
+}
+
+# Download and update Bambu Studio printer access codes
 $ConfPath = "$($env:APPDATA)\BambuStudio\BambuStudio.conf"
 $CodePath = 'codes.txt'
 wget https://github.com/democat3457/PrusaProfiles/raw/refs/heads/main/bambu/codes.txt -OutFile $CodePath
-(Get-Content $ConfPath -Raw).replace("}`r`n}", "},`r`n$((Get-Content $CodePath -Raw))`r`n}") | Set-Content $ConfPath
+if (Test-Path $ConfPath -PathType Leaf) {
+    (Get-Content $ConfPath -Raw).replace("}`r`n}", "},`r`n$((Get-Content $CodePath -Raw))`r`n}") | Set-Content $ConfPath
+    Write-Output "Bambu Studio printer installation complete!"
+}
+else {
+    # Should never happen
+    #New-Item $ConfPath -ItemType File -Force -Value "{`r`n$((Get-Content $CodePath -Raw))`r`n}"
+    Write-Host "Bambu Studio config folder failed to generate! Open Bambu Studio, go through the setup wizard, close the window afterwards, and re-run the script." -BackgroundColor Black -ForegroundColor Red
+}
 rm $CodePath
+pause "Press any key to close..."
